% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscal_year.R
\name{fiscal_year}
\alias{fiscal_year}
\title{Convert dates to fiscal years}
\usage{
fiscal_year(x, start = 4, label = FALSE)
}
\arguments{
\item{x}{A date vector or a vector of class \code{"yearmon"} from the \code{zoo} package.}

\item{start}{An integer between 1 and 12 representing the month on which the fiscal
begins. The default gives April.}

\item{label}{Should labels be returned instead of numeric years? If \code{label = TRUE}
then the out will be a character vector of labels for the appropriate fiscal years,
e.g. "2017/18".}
}
\value{
By default, a numeric vector of fiscal years or, if \code{label = TRUE}, a
  character vector of fiscal years.
}
\description{
Convert a vector of dates or monthly data to the enclosing fiscal year. There are
options to change the start of the fiscal year and choose from either numeric or label
output.
}
\examples{
x <- seq.Date(as.Date("2012-01-01"), as.Date("2017-01-01"), by = "month")

fiscal_year(x)             # April is the default starting month
fiscal_year(x, start = 7)  # But this can be changed

fiscal_year(x, label = TRUE)  # Label output may be attractive when publishing results

}
