% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsaver.R
\name{ggsaver}
\alias{ggsaver}
\title{Helpful defaults for saving ggplot plots}
\usage{
ggsaver(filename, plot = last_plot(), path = NULL, page_size = "A4",
  portrait = FALSE, dpi = 300)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{path}{Path to save plot to (combined with filename).}

\item{page_size}{Page size to use. Must be one of "A0" through to "A10".}

\item{portrait}{The orientation of the page.}

\item{dpi}{Plot resolution.}
}
\description{
This is a wrapper around \code{ggplot2::ggsave()} which allows one to easily save plots
to the commonly used paper sizes of the ISO 216 A standard. "A4" is the default but
the full range - "A0" to "A10" - can be specified.
}
