% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{A binary operator for NOT IN}
\usage{
x \%not_in\% y
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}
}
\value{
A logical vector of the corresponding matches.
}
\description{
Test whether the elements of \code{x} are contained within \code{y}. It is simply a
wrapper around \code{!(x \%in\% y)} but intended as a more readable alternative.
}
